Extract apply_watcher.py from database.py
